bpf: udp: Avoid socket skips during iteration
As it stands today, it's entirely possible to skip a UDP socket during
iteration. The current implementation of bpf_iter_udp_realloc_batch()
uses iter->offset to track progress through a bucket which is a measure
of the number of matching sockets from the current bucket that have
been seen or processed by the iterator. On subsequent calls to
bpf_iter_udp_realloc_batch(), if the current bucket has unprocessed
items, the loop skips at least iter->offset matching items in the bucket
before adding items from that bucket to the next batch. The intent is to
skip any items we've already seen, and there seems to be an underlying
assumption that iter->offset is an accurate measure of "things already
seen". There are, however, a variety of scenarios where the underlying
bucket changes between calls to bpf_iter_udp_realloc_batch(), breaking
that assumption and leading to either repeated or skipped sockets. Two
such scenarios are illustrated below.

Skip A Socket
+------+--------------------+--------------+---------------+
| Time | Event              | Bucket State | Bucket Offset |
+------+--------------------+--------------+---------------+
| 1    | read(iter_fd) -> A | A->B->C->D   | 1             |
| 2    | close(A)           | B->C->D      | 1             |
| 3    | read(iter_fd) -> C | B->C->D      | 2             |
| 4    | read(iter_fd) -> D | B->C->D      | 3             |
| 5    | read(iter_fd) -> 0 | B->C->D      | -             |
+------+--------------------+--------------+---------------+

Iteration sees these buckets: [A, C, D]
B is skipped!

Repeat A Socket
+------+--------------------+---------------+---------------+
| Time | Event              | Bucket State  | Bucket Offset |
+------+--------------------+---------------+---------------+
| 1    | read(iter_fd) -> A | A->B->C->D    | 1             |
| 2    | connect(E)         | E->A->B->C->D | 1             |
| 3    | read(iter_fd) -> A | E->A->B->C->D | 2             |
| 3    | read(iter_fd) -> B | E->A->B->C->D | 3             |
| 3    | read(iter_fd) -> C | E->A->B->C->D | 4             |
| 4    | read(iter_fd) -> D | E->A->B->C->D | 5             |
| 5    | read(iter_fd) -> 0 | E->A->B->C->D | -             |
+------+--------------------+---------------+---------------+

Iteration sees these buckets: [A, A, B, C, D]
A is repeated!

There's some ambiguity over exact iterator semantics. Currently,
semantics are neither at-most-once, nor at-least-once, nor
exactly-once as it relates to the set of sockets that exist at the time
iteration begins. Repeating a socket during iteration is perhaps less
problematic than skipping it altogether as long as the BPF program is
aware that duplicates are possible. However, in an ideal world, we
should process each socket (or other object) exactly once. This is
easier said than done considering these two constraints:

1) Despite batch resize attempts inside bpf_iter_udp_realloc_batch(),
   we have to deal with the possibility that our batch size cannot
   contain all items in a bucket at once and process a single bucket in
   multiple iterations/chunks.
2) We cannot hold a lock on the bucket (hslot2) between iterations,
   meaning that the structure can change in lots of interesting ways
   as illustrated by the two examples above.

Can we achieve exactly-once semantics for iteration over UDP sockets
even in the face of concurrent additions or removals to the current
bucket? If we ignore the possibility of signed 64 bit rollover, then
yes. This patch replaces the current offset-based scheme for progress
tracking with a scheme based on a monotonically increasing version
associated with each bucket. It works as follows:

* Assign index numbers on sockets in the bucket's linked list such that
  they are monotonically increasing as you read from the head to tail.

  * Every time a socket is added to a bucket, increment the bucket's
    version number (hslot2->ver).
  * If the socket is being added to the head of the bucket's linked
    list, set udp_sk(sk)->ver to -1*hslot2->ver.
  * If the socket is being added to the tail of the bucket's linked
    list, set udp_sk(sk)->ver to hslot2->ver.

  Ex: append_head(C), append_head(B), append_tail(D), append_head(A),
      append_tail(E) results in the following state.

      A -> B -> C -> D -> E
     -4   -2   -1    3    5
* As we iterate through a bucket keep track of the last index number
  we've seen for that bucket (iter->prev_idx).
* On calls to bpf_iter_udp_batch(), if we didn't finish reading all
  items in the current bucket last time, skip until we find a socket
  whose index is greater than iter->prev_idx.

Since we always iterate from head to tail and indexes are always
increasing in that direction we can be sure that any socket whose index
is greater than iter->prev_idx has not yet been seen. Any socket whose
index is less than iter->prev_idx has either been seen before or was
added after we began iterating over that bucket. In either case, it's
safe to skip them.