bpf: Add reference count to struct bpf_shtab_elem
Prepare for the next patch by using reference counting to decide when to
free a struct bpf_shtab_elem. Currently, RCU ensures that elements
are not freed while an iterator is handling them. However, the next
patch changes the socket hash iterator algorithm so that bucket elements
are handled outside an RCU read-side critical section making this
insufficient.

Signed-off-by: Jordan Rife <jordan@jrife.io>