selftests/bpf: Socket map + sockops insert and destroy
Use a sockops program to automatically insert sockets into a socket map
and socket hash and use BPF iterators with key prefix bucketing and
filtering to destroy the set of sockets connected to the same remote
port regardless of protocol. This test wraps things up by demonstrating
the desired end to end flow and showing how all the pieces are meant to
fit together.

Signed-off-by: Jordan Rife <jordan@jrife.io>