bpf: Introduce BPF_SOCK_OPS_UDP_CONNECTED_CB
Add the BPF_SOCK_OPS_UDP_CONNECTED_CB callback as a sockops hook where
connected UDP sockets can be inserted into a socket map. This is
invoked on calls to connect() for UDP sockets right after the socket is
hashed. Together with the next patch, this provides the missing piece
allowing us to fully manage the contents of a socket hash in an
environment where we want to keep track of all UDP and TCP sockets
connected to some backend.

is_locked_tcp_sock was recently introduced in [1] to prevent access to
TCP-specific socket fields in contexts where the socket lock isn't held.
This patch extends the use of this field to prevent access to these
fields in UDP socket contexts.

Note: Technically, there should be nothing preventing the use of
      bpf_sock_ops_setsockopt() and bpf_sock_ops_getsockopt() in this
      context, but I've avoided removing the is_locked_tcp_sock_ops()
      guard from these helpers for now to keep the changes in this patch
      series more focused.

[1]: https://lore.kernel.org/all/20250220072940.99994-4-kerneljasonxing@gmail.com/

Signed-off-by: Jordan Rife <jordan@jrife.io>