bpf: Hold socket lock in socket hash iterator
bpf_sock_destroy() must be invoked from a context where the socket lock
is held, but we cannot simply call lock_sock() inside
sock_hash_seq_show(), since it's inside an RCU read side critical
section and lock_sock() may sleep. We also don't want to hold the bucket
lock while running sock_hash_seq_show(), since the BPF program may
itself call map_update_elem() on the same socket hash, acquiring the
same bucket lock and creating a deadlock.

TCP and UDP socket iterators use a batching algorithm to decouple
reading the current bucket's contents and running the BPF iterator
program for each element in the bucket. This enables
sock_hash_seq_show() to acquire the socket lock and lets helpers like
bpf_sock_destroy() run safely. One concern with adopting a similar
algorithm here is that with later patches in the series bucket sizes can
grow arbitrarily large, or at least as large as max_entries for the map.
Naively adopting the same approach risks needing to allocate batches at
least this large to cover the largest bucket size in the map. This could
in theory be mitigated by placing an upper bound on our batch size and
processing a bucket in multiple chunks, but processing in chunks without
a reliable way to track progress through the bucket may lead to skipped
or repeated elements as described in [1]. This could be solved with an
indexing scheme like that described in [2] that associates a
monotonically increasing index to new elements added to the head of the
bucket, but doing so requires an extra 8 bytes to be added to each
element. Not to mention that processing in multiple chunks requires that
we seek to our last position multiple times, making iteration over a large
bucket less efficient.

This patch seeks to do better, requiring no batching and processing all
elements in a bucket in linear time, doing away with the need to seek to
where we left off.

This also fixes a latent bug in the original logic. Before,
sock_hash_seq_start() always called sock_hash_seq_find_next() with
prev_elem set to NULL, forcing iteration to start at the first element
of the current bucket. This logic works under the assumption that
sock_hash_seq_start() is only ever called once for iteration over the
socket hash or that no bucket has more than one element; however, when
using bpf_seq_write sock_hash_seq_start() and sock_hash_seq_stop() may
be called several times as a series of read() calls are made by
userspace, and it may be necessary to resume iteration in the middle of
a bucket. As is, if iteration tries to resume in a bucket with more than
one element it gets stuck, since there is no way to make progress. The
new logic takes such edge cases into account.

[1]: https://lore.kernel.org/bpf/Z_xQhm4aLW9UBykJ@t14/
[2]: https://lore.kernel.org/bpf/20250313233615.2329869-1-jrife@google.com/

Signed-off-by: Jordan Rife <jordan@jrife.io>