bpf: Handle bucket resume scenarios in socket hash iterators
Currently, sock_hash_seq_start() always calls sock_hash_seq_find_next()
with prev_elem set to NULL, forcing iteration to start at the first
element of the current bucket. This logic works under the assumption
that sock_hash_seq_start() is only ever called once for iteration over
the socket hash or that no bucket has more than one element; however,
when using bpf_seq_write sock_hash_seq_start() and sock_hash_seq_stop()
may be called several times as a series of read() calls are made by
userspace, and it may be necessary to resume iteration in the middle of
a bucket. As is, if iteration tries to resume in a bucket with more than
one element it gets stuck, since there is no way to make progress. In
practice bucket collisions are unlikely but not impossible. To handle
edge cases like this, there needs to be a way to keep track of progress
through a bucket. This is complicated by the fact that map elements can
be deleted, updated, or added in arbitrary ways in between iteration
cycles.

One option is to use reference counting to make sure that the current
element and its descendants are not freed even outside an RCU read-side
critical section and even if they're unlinked from the bucket in the
meantime. A drawback of this approach is that it adds an extra eight
bytes for the reference count to each bpf_shtab_elem.

Another option is to take a snapshot of the bucket contents up front
similar to the way UDP and TCP socket iterators work and iterate through
this snapshot instead of the underlying list. It would be unsafe to
reference the original bpf_shtab_elem outside of an RCU read-side
critical section, since it might have been freed, so the snapshot would
consist of a list of unique (key, socket) pairs. To avoid presenting an
inconsistent view of the map in case elements were deleted in between
iteration cycles, each time a (key, socket) pair is to be shown the
iterator would check the real bucket to make sure that element still
exists. A drawback of this approach is that it requires extra memory
allocations and copies during iteration and scanning the bucket list to
check for the existence of (key, socket) pairs could also add extra
work.

Signed-off-by: Jordan Rife <jordan@jrife.io>